# perf(runtime): read entire files in single ops