Pending on https://github.com/denoland/deno/pull/14221

Sync benchmarks:
```
hyperfine '../deno/target/release/deno run -A --no-check test.js' 'node test.js' 'deno run -A --no-check test.js'
Benchmark 1: ../deno/target/release/deno run -A --no-check test.js
  Time (mean ± σ):     163.3 ms ±   2.4 ms    [User: 11.2 ms, System: 152.9 ms]
  Range (min … max):   160.3 ms … 169.3 ms    18 runs

Benchmark 2: node test.js
  Time (mean ± σ):     171.6 ms ±   1.7 ms    [User: 19.5 ms, System: 151.7 ms]
  Range (min … max):   169.5 ms … 175.7 ms    17 runs

Benchmark 3: deno run -A --no-check test.js
  Time (mean ± σ):     172.4 ms ±   1.3 ms    [User: 19.7 ms, System: 155.7 ms]
  Range (min … max):   169.9 ms … 174.4 ms    17 runs

Summary
  '../deno/target/release/deno run -A --no-check test.js' ran
    1.05 ± 0.02 times faster than 'node test.js'
    1.06 ± 0.02 times faster than 'deno run -A --no-check test.js'
```

Async benchmarks:
```
hyperfine '../deno/target/release/deno run -A --no-check test_async.mjs' 'node test_async.mjs' 'deno run -A --no-check test_async.mjs'
Benchmark 1: ../deno/target/release/deno run -A --no-check test_async.mjs
  Time (mean ± σ):     308.3 ms ±   3.3 ms    [User: 11.6 ms, System: 297.8 ms]
  Range (min … max):   305.1 ms … 316.9 ms    10 runs

Benchmark 2: node test_async.mjs
  Time (mean ± σ):     363.8 ms ±   1.0 ms    [User: 40.8 ms, System: 333.1 ms]
  Range (min … max):   362.3 ms … 366.1 ms    10 runs

Benchmark 3: deno run -A --no-check test_async.mjs
  Time (mean ± σ):     672.7 ms ±   5.4 ms    [User: 251.3 ms, System: 453.0 ms]
  Range (min … max):   666.1 ms … 682.2 ms    10 runs

Summary
  '../deno/target/release/deno run -A --no-check test_async.mjs' ran
    1.18 ± 0.01 times faster than 'node test_async.mjs'
    2.18 ± 0.03 times faster than 'deno run -A --no-check test_async.mjs'
```